week4/4_1a_data_2_gcs/requirements.txt
asttokens==2.4.1
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
comm==0.2.1
cramjam==2.8.2
debugpy==1.8.1
decorator==5.1.1
exceptiongroup==1.2.0
executing==2.0.1
fastparquet==2024.2.0
fsspec==2024.2.0
google-api-core==2.17.1
google-auth==2.28.1
google-cloud-core==2.4.1
google-cloud-storage==2.15.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.62.0
idna==3.6
ipykernel==6.29.3
ipython==8.22.2
jedi==0.19.1
jupyter_client==8.6.0
jupyter_core==5.7.1
matplotlib-inline==0.1.6
nest-asyncio==1.6.0
numpy==1.26.4
packaging==23.2
pandas==2.2.1
parso==0.8.3
pexpect==4.9.0
platformdirs==4.2.0
prompt-toolkit==3.0.43
protobuf==4.25.3
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==15.0.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
Pygments==2.17.2
python-dateutil==2.9.0.post0
pytz==2024.1
pyzmq==25.1.2
requests==2.31.0
rsa==4.9
six==1.16.0
stack-data==0.6.3
tornado==6.4
traitlets==5.14.1
tzdata==2024.1
urllib3==2.2.1
wcwidth==0.2.13
